Game Radar Complete
Added funcionality for users to save items to a list so they can keep them on their radar. Component can be managed from it's own route via the game radar drop down link and is also available beneath the shopping cart for quick checkout. Items can be added to a user's radar from the item details page or dropped from the shopping cart to save the item for later.